requirements.txt
requests
python-dotenv
alpha-vantage
sec-edgar-downloader
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
